pyproject.toml
[project]
name = "sample-langchain"
version = "1.0.0"
description = "A sample project for demonstrating LangChain capabilities"

authors = [
    {name = "Ron Webb",email = "ron@ronella.xyz"}
]

readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "transformers (>=4.50.1,<5.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "mcp (>=1.10.1,<2.0.0)",
    "langchain-mcp-adapters (>=0.1.8,<0.2.0)"
]

[tool.poetry.dependencies]
torch = {version = "*", source = "pytorch-cuda"}
torchvision = {version = "*", source = "pytorch-cuda"}
torchaudio = {version = "*", source = "pytorch-cuda"}

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.7"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu126"
priority = "explicit"

[tool.poetry]
name = "sample-langchain"
version = "1.0.0"
description = "A sample project for demonstrating LangChain capabilities"
authors = ["Ron Webb <ron@ronella.xyz>"]
readme = "README.md"
packages = [{include = "sample"}]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:Failing to pass a value to the 'type_params' parameter of 'typing.ForwardRef._evaluate' is deprecated:DeprecationWarning:pydantic.v1.typing",
]